refactor: 초대 판정 규칙을 목록 조작에서 떼어냄 #88
invites 270 → 179. 짝 규칙·중복 판정은 invite-rules로 옮겼다 — 화면과 목록이
각자 판정하면 경고는 뜨는데 저장은 되는 줄이 생긴다.
